backend/requirements.txt
asgiref==3.2.7
astroid==2.4.1
beautifulsoup4==4.9.1
boto3==1.13.21
botocore==1.16.21
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.6
django-cors-headers==3.3.0
django-filter==2.2.0
django-location-field==2.1.0
django-mysql==3.5.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.16
dotty-dict==1.2.1
drf-yasg==1.17.1
gitdb==4.0.5
GitPython==3.1.3
idna==2.9
importlib-metadata==1.6.0
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.18.4
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
pandas==1.0.4
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
quotes==0.0.3
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
setuptools-scm==4.1.2
simplejson==3.17.0
six==1.15.0
smmap==3.0.4
social-auth-app-django==3.4.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.12.1
zipp==3.1.0
